% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syssvm.R
\name{syssvm}
\alias{syssvm}
\title{Run sysSVM}
\usage{
syssvm(input.file = NULL, output.dir = NULL, exclude.features = NULL,
  models = NULL, scaling.factors = NULL)
}
\arguments{
\item{object}{an object, whose class determines}
}
\description{
\code{syssvm} uses e1071 to run an one-class SVM
}
